+++ ProjectEuler/ProjectEuler/Utils/EulerUtils.h
@@ -10,7 +10,7 @@ namespace Problems{
 		static int findNextPrime(List<int> * &primes, __int64 target = -1);
 		static List<int> findFactors(__int64 target, List<int> * primes = &List<int>());
 		static bool isPrime(__int64 num, int accuracy=5);
-		static int modularPow(__int64 base, int exponent, int modulus);
+		static int modularPow(__int64 base, __int64 exponent, __int64 modulus);
 	};
 }
 #endif
